Added new 'with'-methods that accept a named getter.
Where a named getter is an Attribute.

I found that the self-description mechanism, though it worked for most of the cases did not always yield helpful descriptions.
Because the names of the called methods are not always helpful.
So, it's not worth all that reflective tracking that's going on.
Kill your darlings, just give a name to the getters.

Also overloaded the new 'with' method such that it can take a plain Object and a Matcher.
(Because intuitively, you get used to type ".with..." for the autocompletion.)
That method delegates to typesafe methods 'withValue' and 'withMatching'.
Tried to achieve a minimum of type-mismatch tolerance.

The old method will be deprecated some time soon.