Deprecated self-describing 'with' and 'having'.
Self-describing is cumbersome. Having to set a system property upfront is too painful.
Because when turned on it's way too slow.
It introduces many dependencies and a lot of reflective processing.
Which may cause trouble on migration to Java 9.
And the generated descriptions are often useless when the called methods don't have telling names.

Conclusion: Go away.